fix: wrong bounding coordinates for highlights
This should fix #10400.

After a multiline text selection in PDF, the first rectangle in sel-range is a incorrect "noise", not relevant to the selection range. The underlying cause is not clear to me. Note that it's a universal problem when doing mutliline highlights, even though most of the time it does not have any effect on user experience.   

This commit is a working hotfix. And very possibly there would be a better solution after looking into the update mechanism of the variable `sel-state` used here. But I have no enough experience to dive into this.